requirements.txt
coinbase_advanced_py==1.8.2
matplotlib==3.10.8
